Restructured project directories, moved a renamed components and entities.
